orchestrator/requirements.txt
feedparser
requests
beautifulsoup4
sentence-transformers
faiss-cpu
langgraph
pydub
python-dateutil
tqdm
fastapi
uvicorn[standard]
google-cloud-texttospeech
google-cloud-storage
google-cloud-secret-manager
numpy
scikit-learn
tldextract
ujson
openai
tenacity
trafilatura
readability-lxml
newspaper3k
goose3
aiohttp
langdetect
pyrate-limiter
rank-bm25